requirements.txt
# JMS-Sync 依赖库

# 基础库
requests>=2.28.0
pyyaml>=6.0
drf-httpsig>=1.0.0
httpsig>=1.3.0
requests-http-signature>=0.2.0,<1.0.0

# 阿里云SDK
alibabacloud-ecs20140526>=6.0.0
alibabacloud-tea-openapi>=0.3.7
alibabacloud-tea-util>=0.3.7

# 华为云SDK
huaweicloudsdkcore>=3.1.77
huaweicloudsdkecs>=3.1.77

# 工具库
python-dateutil>=2.8.2
pytz>=2021.3
jmespath>=0.10.0
six>=1.16.0

# 加密相关
cryptography>=36.0.1
pycryptodome>=3.14.1
cffi>=1.15.0
pycparser>=2.21

# 测试相关
pytest>=7.0.0
pytest-cov>=3.0.0
mock>=4.0.3

# 开发工具
black>=22.1.0
isort>=5.10.1
mypy>=0.931
flake8>=4.0.1